+++ backend/libraries/utils/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod base64;
 pub mod canister;
 pub mod canister_event_sync_queue;
 pub mod canister_timers;